Fix missing instance error
Seems like the instance resolution algorithm has changed?  Need to bring
the instance for the single inhabitant of the unit type into scope in
order to make the instance of `FromNat` for `Level` compile.  So we need
to open the unit module within `FromNat` publically.

The whole project still doesn’t compile, though.  More coming.